refactor(guild): split monolithic handlers.rs into sub-domain files
