+++ server/src/guild/settings.rs
@@ -1,4 +1,4 @@
-//! Guild Management Handlers
+//! Guild settings, usage stats, channels list, and banner upload handlers.
 
 use axum::extract::{Multipart, Path, State};
 use axum::http::StatusCode;
@@ -7,14 +7,10 @@ use chrono::Utc;
 use serde::{Deserialize, Serialize};
 use sqlx::QueryBuilder;
 use uuid::Uuid;
-use validator::Validate;
 
 use super::error::GuildError;
 use super::queries::{core as core_q, limits};
-use super::types::{
-    CreateGuildRequest, Guild, GuildCommandInfo, GuildMember, GuildSettings, GuildWithMemberCount,
-    UpdateGuildRequest, UpdateGuildSettingsRequest,
-};
+use super::types::{Guild, GuildSettings, UpdateGuildSettingsRequest};
 use crate::api::AppState;
 use crate::auth::AuthUser;
 use crate::db::{self, ChannelType};
@@ -41,21 +37,6 @@ pub struct ChannelWithUnread {
     pub last_message_id: Option<Uuid>,
 }
 
-/// A bot installed in a guild.
-#[derive(Debug, Serialize, sqlx::FromRow, utoipa::ToSchema)]
-pub struct InstalledBot {
-    pub application_id: Uuid,
-    pub bot_user_id: Uuid,
-    pub name: String,
-    pub description: Option<String>,
-    pub installed_by: Uuid,
-    pub installed_at: chrono::DateTime<chrono::Utc>,
-}
-
-// ============================================================================
-// Request Types
-// ============================================================================
-
 /// Position specification for a channel in reorder request.
 #[derive(Debug, Deserialize, utoipa::ToSchema)]
 pub struct ChannelPosition {
@@ -71,434 +52,29 @@ pub struct ReorderChannelsRequest {
     pub channels: Vec<ChannelPosition>,
 }
 
+/// A single resource usage metric (current count vs limit).
+#[derive(Debug, Serialize, utoipa::ToSchema)]
+pub struct UsageStat {
+    pub current: i64,
+    pub limit: i64,
+}
+
+/// Guild resource usage statistics.
+#[derive(Debug, Serialize, utoipa::ToSchema)]
+pub struct GuildUsageStats {
+    pub guild_id: Uuid,
+    pub plan: String,
+    pub members: UsageStat,
+    pub channels: UsageStat,
+    pub roles: UsageStat,
+    pub emojis: UsageStat,
+    pub bots: UsageStat,
+    pub pages: UsageStat,
+}
+
 // ============================================================================
 // Handlers
 // ============================================================================
-
-/// Create a new guild
-#[utoipa::path(
-    post,
-    path = "/api/guilds",
-    tag = "guilds",
-    request_body = CreateGuildRequest,
-    responses((status = 200, body = Guild)),
-    security(("bearer_auth" = []))
-)]
-#[tracing::instrument(skip(state))]
-pub async fn create_guild(
-    State(state): State<AppState>,
-    auth: AuthUser,
-    Json(body): Json<CreateGuildRequest>,
-) -> Result<Json<Guild>, GuildError> {
-    // Validate request
-    body.validate()
-        .map_err(|e| GuildError::Validation(crate::validation::format_validation_errors(&e)))?;
-
-    // Validate tags if provided
-    if let Some(ref tags) = body.tags {
-        if tags.len() > 5 {
-            return Err(GuildError::Validation("Maximum 5 tags allowed".to_string()));
-        }
-        for tag in tags {
-            if tag.len() < 2 || tag.len() > 32 {
-                return Err(GuildError::Validation(
-                    "Each tag must be 2-32 characters".to_string(),
-                ));
-            }
-            if !TAG_REGEX.is_match(tag) {
-                return Err(GuildError::Validation(
-                    "Tags may only contain letters, numbers, and hyphens".to_string(),
-                ));
-            }
-        }
-    }
-
-    // Validate banner_url if provided
-    if let Some(ref url) = body.banner_url {
-        if !url.is_empty() {
-            if url.len() > 2048 {
-                return Err(GuildError::Validation(
-                    "Banner URL too long (max 2048 characters)".to_string(),
-                ));
-            }
-            if !url.starts_with("https://") {
-                return Err(GuildError::Validation(
-                    "Banner URL must use HTTPS".to_string(),
-                ));
-            }
-        }
-    }
-
-    let mut tx = state.db.begin().await?;
-
-    // Serialize guild creation per owner to enforce strict user guild limits.
-    core_q::lock_guild_create_for_user(&mut tx, auth.id).await?;
-
-    // Check guild creation limit
-    let owned_count = core_q::count_user_owned_guilds_tx(&mut tx, auth.id).await?;
-    if owned_count >= state.config.max_guilds_per_user {
-        return Err(GuildError::LimitExceeded(format!(
-            "Maximum number of guilds reached ({})",
-            state.config.max_guilds_per_user
-        )));
-    }
-
-    // Prepare discovery fields
-    let discoverable = body.discoverable.unwrap_or(false);
-    let tags: Vec<String> = body
-        .tags
-        .map(|t| t.into_iter().map(|s| s.to_lowercase()).collect())
-        .unwrap_or_default();
-    let banner_url: Option<String> =
-        body.banner_url
-            .and_then(|u| if u.is_empty() { None } else { Some(u) });
-
-    // Insert guild with discovery fields
-    let guild_id = Uuid::now_v7();
-    let guild = core_q::insert_guild(
-        &mut tx,
-        guild_id,
-        &body.name,
-        auth.id,
-        &body.description,
-        discoverable,
-        &tags,
-        &banner_url,
-    )
-    .await?;
-
-    // Add owner as member
-    core_q::insert_owner_member(&mut tx, guild_id, auth.id).await?;
-
-    // Create default @everyone role with sensible default permissions
-    core_q::insert_default_everyone_role(
-        &mut tx,
-        guild_id,
-        crate::permissions::GuildPermissions::EVERYONE_DEFAULT.bits() as i64,
-    )
-    .await?;
-
-    tx.commit().await?;
-
-    Ok(Json(guild))
-}
-
-/// List guilds for the current user with member counts
-#[utoipa::path(
-    get,
-    path = "/api/guilds",
-    tag = "guilds",
-    responses((status = 200, body = Vec<GuildWithMemberCount>)),
-    security(("bearer_auth" = []))
-)]
-#[tracing::instrument(skip(state))]
-pub async fn list_guilds(
-    State(state): State<AppState>,
-    auth: AuthUser,
-) -> Result<Json<Vec<GuildWithMemberCount>>, GuildError> {
-    // Query guilds with member count in a single query
-    let rows = core_q::list_guilds_with_member_count(&state.db, auth.id).await?;
-
-    let guilds = rows
-        .into_iter()
-        .map(
-            |(
-                id,
-                name,
-                owner_id,
-                icon_url,
-                description,
-                threads_enabled,
-                discoverable,
-                tags,
-                banner_url,
-                plan,
-                created_at,
-                member_count,
-            )| {
-                GuildWithMemberCount {
-                    guild: Guild {
-                        id,
-                        name,
-                        owner_id,
-                        icon_url,
-                        description,
-                        threads_enabled,
-                        discoverable,
-                        tags,
-                        banner_url,
-                        plan,
-                        created_at,
-                    },
-                    member_count,
-                }
-            },
-        )
-        .collect();
-
-    Ok(Json(guilds))
-}
-
-/// Get guild details
-#[utoipa::path(
-    get,
-    path = "/api/guilds/{id}",
-    tag = "guilds",
-    params(("id" = Uuid, Path, description = "Guild ID")),
-    responses((status = 200, body = Guild)),
-    security(("bearer_auth" = []))
-)]
-#[tracing::instrument(skip(state))]
-pub async fn get_guild(
-    State(state): State<AppState>,
-    auth: AuthUser,
-    Path(guild_id): Path<Uuid>,
-) -> Result<Json<Guild>, GuildError> {
-    // Verify membership
-    let is_member = db::is_guild_member(&state.db, guild_id, auth.id).await?;
-    if !is_member {
-        return Err(GuildError::Forbidden);
-    }
-
-    let guild = core_q::fetch_guild(&state.db, guild_id).await?;
-
-    Ok(Json(guild))
-}
-
-/// Update guild
-#[utoipa::path(
-    patch,
-    path = "/api/guilds/{id}",
-    tag = "guilds",
-    params(("id" = Uuid, Path, description = "Guild ID")),
-    request_body = UpdateGuildRequest,
-    responses((status = 200, body = Guild)),
-    security(("bearer_auth" = []))
-)]
-#[tracing::instrument(skip(state))]
-pub async fn update_guild(
-    State(state): State<AppState>,
-    auth: AuthUser,
-    Path(guild_id): Path<Uuid>,
-    Json(body): Json<UpdateGuildRequest>,
-) -> Result<Json<Guild>, GuildError> {
-    // Validate request
-    body.validate()
-        .map_err(|e| GuildError::Validation(crate::validation::format_validation_errors(&e)))?;
-
-    // Verify ownership
-    let owner_id = core_q::fetch_guild_owner(&state.db, guild_id).await?;
-
-    if owner_id != auth.id {
-        return Err(GuildError::Forbidden);
-    }
-
-    // Build dynamic update query
-    let mut has_changes = false;
-    let mut builder = QueryBuilder::new("UPDATE guilds SET ");
-    {
-        let mut sep = builder.separated(", ");
-        if let Some(name) = body.name {
-            sep.push("name = ").push_bind_unseparated(name);
-            has_changes = true;
-        }
-        if let Some(desc) = body.description {
-            sep.push("description = ").push_bind_unseparated(desc);
-            has_changes = true;
-        }
-        if let Some(icon) = body.icon_url {
-            sep.push("icon_url = ").push_bind_unseparated(icon);
-            has_changes = true;
-        }
-    }
-
-    if !has_changes {
-        return get_guild(State(state), auth, Path(guild_id)).await;
-    }
-
-    let updated_guild = core_q::update_guild_dynamic(&state.db, builder, guild_id).await?;
-
-    Ok(Json(updated_guild))
-}
-
-/// Delete guild (owner only)
-#[utoipa::path(
-    delete,
-    path = "/api/guilds/{id}",
-    tag = "guilds",
-    params(("id" = Uuid, Path, description = "Guild ID")),
-    responses((status = 204, description = "Guild deleted")),
-    security(("bearer_auth" = []))
-)]
-#[tracing::instrument(skip(state))]
-pub async fn delete_guild(
-    State(state): State<AppState>,
-    auth: AuthUser,
-    Path(guild_id): Path<Uuid>,
-) -> Result<StatusCode, GuildError> {
-    // Verify ownership
-    let owner_id = core_q::fetch_guild_owner(&state.db, guild_id).await?;
-
-    if owner_id != auth.id {
-        return Err(GuildError::Forbidden);
-    }
-
-    core_q::delete_guild(&state.db, guild_id).await?;
-
-    Ok(StatusCode::NO_CONTENT)
-}
-
-/// Initialize `channel_read_state` for all text channels in a guild.
-/// Sets `last_read_at` to `NOW()` so pre-existing messages don't appear as unread.
-pub(crate) async fn initialize_channel_read_state(
-    db: &sqlx::PgPool,
-    guild_id: Uuid,
-    user_id: Uuid,
-) -> Result<(), GuildError> {
-    core_q::initialize_channel_read_state(db, guild_id, user_id).await
-}
-
-/// Leave guild
-#[utoipa::path(
-    post,
-    path = "/api/guilds/{id}/leave",
-    tag = "guilds",
-    params(("id" = Uuid, Path, description = "Guild ID")),
-    responses((status = 204, description = "Left guild")),
-    security(("bearer_auth" = []))
-)]
-#[tracing::instrument(skip(state))]
-pub async fn leave_guild(
-    State(state): State<AppState>,
-    auth: AuthUser,
-    Path(guild_id): Path<Uuid>,
-) -> Result<StatusCode, GuildError> {
-    // Verify membership
-    let is_member = db::is_guild_member(&state.db, guild_id, auth.id).await?;
-    if !is_member {
-        return Err(GuildError::NotFound);
-    }
-
-    // Check if owner (owners can't leave, must transfer ownership first)
-    let owner_id = core_q::fetch_guild_owner_required(&state.db, guild_id).await?;
-
-    if owner_id == auth.id {
-        return Err(GuildError::Validation(
-            "Guild owner must transfer ownership before leaving".to_string(),
-        ));
-    }
-
-    // Remove membership
-    core_q::delete_guild_member(&state.db, guild_id, auth.id).await?;
-
-    // Dispatch MemberLeft to bot ecosystem (non-blocking)
-    {
-        let db = state.db.clone();
-        let redis = state.redis.clone();
-        let gid = guild_id;
-        let uid = auth.id;
-        tokio::spawn(async move {
-            crate::ws::bot_events::publish_member_left(&db, &redis, gid, uid).await;
-            crate::webhooks::dispatch::dispatch_guild_event(
-                &db,
-                &redis,
-                gid,
-                crate::webhooks::events::BotEventType::MemberLeft,
-                serde_json::json!({ "guild_id": gid, "user_id": uid }),
-            )
-            .await;
-        });
-    }
-
-    Ok(StatusCode::NO_CONTENT)
-}
-
-/// List guild members
-#[utoipa::path(
-    get,
-    path = "/api/guilds/{id}/members",
-    tag = "guilds",
-    params(("id" = Uuid, Path, description = "Guild ID")),
-    responses((status = 200, body = Vec<GuildMember>)),
-    security(("bearer_auth" = []))
-)]
-#[tracing::instrument(skip(state))]
-pub async fn list_members(
-    State(state): State<AppState>,
-    auth: AuthUser,
-    Path(guild_id): Path<Uuid>,
-) -> Result<Json<Vec<GuildMember>>, GuildError> {
-    // Verify membership
-    let is_member = db::is_guild_member(&state.db, guild_id, auth.id).await?;
-    if !is_member {
-        return Err(GuildError::Forbidden);
-    }
-
-    let members = core_q::list_guild_members(&state.db, guild_id).await?;
-
-    Ok(Json(members))
-}
-
-/// Kick a member from guild (owner only)
-#[utoipa::path(
-    delete,
-    path = "/api/guilds/{id}/members/{user_id}",
-    tag = "guilds",
-    params(
-        ("id" = Uuid, Path, description = "Guild ID"),
-        ("user_id" = Uuid, Path, description = "User ID")
-    ),
-    responses((status = 204, description = "Member kicked")),
-    security(("bearer_auth" = []))
-)]
-#[tracing::instrument(skip(state))]
-pub async fn kick_member(
-    State(state): State<AppState>,
-    auth: AuthUser,
-    Path((guild_id, user_id)): Path<(Uuid, Uuid)>,
-) -> Result<StatusCode, GuildError> {
-    // Verify ownership
-    let owner_id = core_q::fetch_guild_owner(&state.db, guild_id).await?;
-
-    if owner_id != auth.id {
-        return Err(GuildError::Forbidden);
-    }
-
-    // Cannot kick yourself (owner)
-    if user_id == auth.id {
-        return Err(GuildError::Validation(
-            "Cannot kick yourself from the guild".to_string(),
-        ));
-    }
-
-    // Remove membership
-    let rows_affected = core_q::delete_guild_member(&state.db, guild_id, user_id).await?;
-
-    if rows_affected == 0 {
-        return Err(GuildError::NotFound);
-    }
-
-    // Dispatch MemberLeft to bot ecosystem (non-blocking)
-    {
-        let db = state.db.clone();
-        let redis = state.redis.clone();
-        let gid = guild_id;
-        let uid = user_id;
-        tokio::spawn(async move {
-            crate::ws::bot_events::publish_member_left(&db, &redis, gid, uid).await;
-            crate::webhooks::dispatch::dispatch_guild_event(
-                &db,
-                &redis,
-                gid,
-                crate::webhooks::events::BotEventType::MemberLeft,
-                serde_json::json!({ "guild_id": gid, "user_id": uid }),
-            )
-            .await;
-        });
-    }
-
-    Ok(StatusCode::NO_CONTENT)
-}
 
 /// List guild channels with unread counts
 #[utoipa::path(
@@ -651,194 +227,6 @@ pub async fn reorder_channels(
     tx.commit().await?;
 
     Ok(StatusCode::NO_CONTENT)
-}
-
-/// Install a bot into a guild.
-///
-/// `POST /api/guilds/:guild_id/bots/:bot_id/add`
-#[utoipa::path(
-    post,
-    path = "/api/guilds/{id}/bots/{bot_id}/add",
-    tag = "guilds",
-    params(
-        ("id" = Uuid, Path, description = "Guild ID"),
-        ("bot_id" = Uuid, Path, description = "Bot user ID")
-    ),
-    responses((status = 204, description = "Bot added")),
-    security(("bearer_auth" = []))
-)]
-#[tracing::instrument(skip(state))]
-pub async fn add_bot_to_guild(
-    State(state): State<AppState>,
-    auth: AuthUser,
-    Path((guild_id, bot_id)): Path<(Uuid, Uuid)>,
-) -> Result<StatusCode, GuildError> {
-    let _ctx =
-        require_guild_permission(&state.db, guild_id, auth.id, GuildPermissions::MANAGE_GUILD)
-            .await
-            .map_err(|e| match e {
-                PermissionError::NotGuildMember => GuildError::Forbidden,
-                other => GuildError::Permission(other),
-            })?;
-
-    // Validate bot exists and is installable (outside lock)
-    if core_q::fetch_bot_user(&state.db, bot_id).await?.is_none() {
-        return Err(GuildError::NotFound);
-    }
-
-    let app = core_q::fetch_bot_application(&state.db, bot_id)
-        .await?
-        .ok_or(GuildError::NotFound)?;
-
-    if !app.public && app.owner_id != auth.id {
-        return Err(GuildError::NotFound);
-    }
-
-    let application_id = app.id;
-
-    // Advisory lock seed 63 = bot_install (see db/mod.rs registry)
-    let mut tx = state.db.begin().await?;
-
-    core_q::lock_bot_install(&mut tx, guild_id).await?;
-
-    let bot_count = core_q::count_guild_bots_tx(&mut tx, guild_id).await?;
-
-    if bot_count >= state.config.max_bots_per_guild {
-        return Err(GuildError::LimitExceeded(format!(
-            "Maximum number of bots per guild reached ({})",
-            state.config.max_bots_per_guild
-        )));
-    }
-
-    core_q::insert_bot_installation(&mut tx, guild_id, application_id, auth.id).await?;
-
-    tx.commit().await?;
-
-    Ok(StatusCode::NO_CONTENT)
-}
-
-/// List bots installed in a guild.
-///
-/// `GET /api/guilds/:guild_id/bots`
-#[utoipa::path(
-    get,
-    path = "/api/guilds/{id}/bots",
-    tag = "guilds",
-    params(("id" = Uuid, Path, description = "Guild ID")),
-    responses((status = 200, body = Vec<InstalledBot>)),
-    security(("bearer_auth" = []))
-)]
-#[tracing::instrument(skip(state))]
-pub async fn list_guild_bots(
-    State(state): State<AppState>,
-    auth: AuthUser,
-    Path(guild_id): Path<Uuid>,
-) -> Result<Json<Vec<InstalledBot>>, GuildError> {
-    // Verify membership
-    let is_member = db::is_guild_member(&state.db, guild_id, auth.id).await?;
-    if !is_member {
-        return Err(GuildError::Forbidden);
-    }
-
-    let bots = core_q::list_guild_bots(&state.db, guild_id).await?;
-
-    Ok(Json(bots))
-}
-
-/// Remove a bot from a guild.
-///
-/// `DELETE /api/guilds/:guild_id/bots/:bot_id`
-#[utoipa::path(
-    delete,
-    path = "/api/guilds/{id}/bots/{bot_id}",
-    tag = "guilds",
-    params(
-        ("id" = Uuid, Path, description = "Guild ID"),
-        ("bot_id" = Uuid, Path, description = "Bot user ID")
-    ),
-    responses((status = 204, description = "Bot removed")),
-    security(("bearer_auth" = []))
-)]
-#[tracing::instrument(skip(state))]
-pub async fn remove_bot_from_guild(
-    State(state): State<AppState>,
-    auth: AuthUser,
-    Path((guild_id, bot_id)): Path<(Uuid, Uuid)>,
-) -> Result<StatusCode, GuildError> {
-    let _ctx =
-        require_guild_permission(&state.db, guild_id, auth.id, GuildPermissions::MANAGE_GUILD)
-            .await
-            .map_err(|e| match e {
-                PermissionError::NotGuildMember => GuildError::Forbidden,
-                other => GuildError::Permission(other),
-            })?;
-
-    // Look up application_id from bot_user_id
-    let application_id = core_q::fetch_bot_application_id(&state.db, bot_id)
-        .await?
-        .ok_or(GuildError::NotFound)?;
-
-    let rows_affected =
-        core_q::delete_bot_installation(&state.db, guild_id, application_id).await?;
-
-    if rows_affected == 0 {
-        return Err(GuildError::NotFound);
-    }
-
-    Ok(StatusCode::NO_CONTENT)
-}
-
-/// List available slash commands in a guild (from installed bots).
-///
-/// Returns both guild-scoped and global commands from all installed bots.
-/// Guild-scoped commands take precedence over global commands with the same name.
-///
-/// `GET /api/guilds/:guild_id/commands`
-#[utoipa::path(
-    get,
-    path = "/api/guilds/{id}/commands",
-    tag = "guilds",
-    params(("id" = Uuid, Path, description = "Guild ID")),
-    responses((status = 200, body = Vec<GuildCommandInfo>)),
-    security(("bearer_auth" = []))
-)]
-#[tracing::instrument(skip(state))]
-pub async fn list_guild_commands(
-    State(state): State<AppState>,
-    auth: AuthUser,
-    Path(guild_id): Path<Uuid>,
-) -> Result<Json<Vec<GuildCommandInfo>>, GuildError> {
-    // Verify membership
-    let is_member = db::is_guild_member(&state.db, guild_id, auth.id).await?;
-    if !is_member {
-        return Err(GuildError::Forbidden);
-    }
-
-    // Return all commands from installed bots (no DISTINCT ON).
-    let rows = core_q::list_guild_slash_commands(&state.db, guild_id).await?;
-
-    // Compute ambiguity: count how many distinct apps provide each command name.
-    let mut name_counts: std::collections::HashMap<String, usize> =
-        std::collections::HashMap::new();
-    for (name, _, _, _) in &rows {
-        *name_counts.entry(name.clone()).or_insert(0) += 1;
-    }
-
-    let result: Vec<GuildCommandInfo> = rows
-        .into_iter()
-        .map(|(name, description, bot_name, application_id)| {
-            let is_ambiguous = name_counts.get(&name).copied().unwrap_or(0) > 1;
-            GuildCommandInfo {
-                name,
-                description,
-                bot_name,
-                application_id,
-                is_ambiguous,
-            }
-        })
-        .collect();
-
-    Ok(Json(result))
 }
 
 /// Mark all text channels in a guild as read.
@@ -1056,30 +444,6 @@ pub async fn dismiss_discovery_prompt(
     core_q::dismiss_discovery_prompt(&state.db, guild_id, auth.id).await?;
 
     Ok(StatusCode::NO_CONTENT)
-}
-
-// ============================================================================
-// Guild Usage Stats
-// ============================================================================
-
-/// A single resource usage metric (current count vs limit).
-#[derive(Debug, Serialize, utoipa::ToSchema)]
-pub struct UsageStat {
-    pub current: i64,
-    pub limit: i64,
-}
-
-/// Guild resource usage statistics.
-#[derive(Debug, Serialize, utoipa::ToSchema)]
-pub struct GuildUsageStats {
-    pub guild_id: Uuid,
-    pub plan: String,
-    pub members: UsageStat,
-    pub channels: UsageStat,
-    pub roles: UsageStat,
-    pub emojis: UsageStat,
-    pub bots: UsageStat,
-    pub pages: UsageStat,
 }
 
 /// Get guild resource usage stats.
