+++ server/src/openapi.rs
@@ -120,24 +120,24 @@ use utoipa::{Modify, OpenApi};
         crate::chat::overrides::set_override,
         crate::chat::overrides::delete_override,
         // Guilds
-        crate::guild::handlers::list_guilds,
-        crate::guild::handlers::create_guild,
-        crate::guild::handlers::get_guild,
-        crate::guild::handlers::update_guild,
-        crate::guild::handlers::delete_guild,
-        crate::guild::handlers::leave_guild,
-        crate::guild::handlers::list_members,
-        crate::guild::handlers::kick_member,
-        crate::guild::handlers::list_channels,
-        crate::guild::handlers::reorder_channels,
-        crate::guild::handlers::mark_all_channels_read,
-        crate::guild::handlers::list_guild_bots,
-        crate::guild::handlers::add_bot_to_guild,
-        crate::guild::handlers::remove_bot_from_guild,
-        crate::guild::handlers::list_guild_commands,
-        crate::guild::handlers::get_guild_settings,
-        crate::guild::handlers::update_guild_settings,
-        crate::guild::handlers::get_guild_usage,
+        crate::guild::core::list_guilds,
+        crate::guild::core::create_guild,
+        crate::guild::core::get_guild,
+        crate::guild::core::update_guild,
+        crate::guild::core::delete_guild,
+        crate::guild::members::leave_guild,
+        crate::guild::members::list_members,
+        crate::guild::members::kick_member,
+        crate::guild::settings::list_channels,
+        crate::guild::settings::reorder_channels,
+        crate::guild::settings::mark_all_channels_read,
+        crate::guild::bots::list_guild_bots,
+        crate::guild::bots::add_bot_to_guild,
+        crate::guild::bots::remove_bot_from_guild,
+        crate::guild::bots::list_guild_commands,
+        crate::guild::settings::get_guild_settings,
+        crate::guild::settings::update_guild_settings,
+        crate::guild::settings::get_guild_usage,
         // Roles
         crate::guild::roles::list_roles,
         crate::guild::roles::create_role,
@@ -424,12 +424,12 @@ use utoipa::{Modify, OpenApi};
         crate::guild::types::GuildSettings,
         crate::guild::types::UpdateGuildSettingsRequest,
         crate::guild::types::GuildCommandInfo,
-        crate::guild::handlers::UsageStat,
-        crate::guild::handlers::GuildUsageStats,
-        crate::guild::handlers::ChannelWithUnread,
-        crate::guild::handlers::InstalledBot,
-        crate::guild::handlers::ChannelPosition,
-        crate::guild::handlers::ReorderChannelsRequest,
+        crate::guild::settings::UsageStat,
+        crate::guild::settings::GuildUsageStats,
+        crate::guild::settings::ChannelWithUnread,
+        crate::guild::bots::InstalledBot,
+        crate::guild::settings::ChannelPosition,
+        crate::guild::settings::ReorderChannelsRequest,
         // Guild - Categories
         crate::guild::categories::Category,
         crate::guild::categories::CreateCategoryRequest,
